VMRestore: allow creating vmrestore even if runstrategy is not Halted
Code had a faulty assumption that Always or RerunOnFailure
means that VM cannot be stopped.
Instead we dont check the RunStrategy or Running field and we look
on the VMI if it exists or not.
(This fix is not ideal since we prefere not to check other
objects in the webhook since it's racy but, this is the
minimal API change possible for a bug fix)
This issue was already handled in for vmsnapshot
process but also applies here: #12515
This fixes:
https://issues.redhat.com/browse/CNV-51071

Signed-off-by: Shelly Kagan <skagan@redhat.com>