fix(virtctl,create): Validate names of disks
Validate the names of disks that are derived from the VM's name or
specified by the user against RFC 1123 DNS label rules to avoid
generating VMs with disk names that will lead to the VM being
rejected upon creation.

Note: The conflicts were due to some cleanups missing in this
branch.

Conflicts:
  pkg/virtctl/create/vm/vm.go
  pkg/virtctl/create/vm/vm_test.go

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>